Introduced LineTransformationOutputStream.Delegating. MavenConsoleAnnotator had neglected to implement flush properly.
(cherry picked from commit 4c6be45e6f1aed49f9eae87be3a98086267fd436)
