+++ core/src/main/java/hudson/console/ConsoleAnnotationOutputStream.java
@@ -49,7 +49,7 @@ import org.kohsuke.stapler.framework.io.WriterOutputStream;
  * @since 1.349
  */
 public class ConsoleAnnotationOutputStream<T> extends LineTransformationOutputStream {
-    private final Writer out;
+    private final Writer out; // not an OutputStream so cannot use LineTransformationOutputStream.Delegating
     private final T context;
     private ConsoleAnnotator<T> ann;
