+++ test/src/test/java/jenkins/tasks/SimpleBuildWrapperTest.java
@@ -187,10 +187,10 @@ public class SimpleBuildWrapperTest {
         private static class UpcaseFilter extends ConsoleLogFilter implements Serializable {
             private static final long serialVersionUID = 1;
             @SuppressWarnings("rawtypes") // inherited
-            @Override public OutputStream decorateLogger(AbstractBuild _ignore, final OutputStream logger) throws IOException, InterruptedException {
-                return new LineTransformationOutputStream() {
+            @Override public OutputStream decorateLogger(AbstractBuild _ignore, OutputStream logger) throws IOException, InterruptedException {
+                return new LineTransformationOutputStream.Delegating(logger) {
                     @Override protected void eol(byte[] b, int len) throws IOException {
-                        logger.write(new String(b, 0, len).toUpperCase(Locale.ROOT).getBytes());
+                        out.write(new String(b, 0, len).toUpperCase(Locale.ROOT).getBytes());
                     }
                 };
             }
